requirements.txt
chromadb==0.4.24
pdfplumber==0.11.0
sentence-transformers==2.7.0    
pytesseract==0.3.10
pdf2image==1.17.0
ollama==0.2.1                   
python-dotenv==1.0.1
tqdm==4.66.2
Pillow==10.3.0
poppler-windows==25.12.0